Moves around some code
